+++ contracts/furion-swap/interfaces/IFurionSwapV2Router.sol
@@ -4,104 +4,90 @@ pragma solidity ^0.8.10;
 
 interface IFurionSwapV2Router {
     function factory() external view returns (address);
-
     function WETH() external view returns (address);
 
     function addLiquidity(
         address _tokenA,
         address _tokenB,
-        uint256 _amountADesired,
-        uint256 _amountBDesired,
-        uint256 _amountAMin,
-        uint256 _amountBMin,
+        uint _amountADesired,
+        uint _amountBDesired,
+        uint _amountAMin,
+        uint _amountBMin,
         address _to,
-        uint256 _deadline
-    )
-        external
-        returns (
-            uint256 _amountA,
-            uint256 _amountB,
-            uint256 _liquidity
-        );
+        uint _deadline
+    ) external returns (uint _amountA, uint _amountB, uint _liquidity);
 
     function addLiquidityETH(
         address _token,
-        uint256 _amountTokenDesired,
-        uint256 _amountTokenMin,
-        uint256 _amountETHMin,
+        uint _amountTokenDesired,
+        uint _amountTokenMin,
+        uint _amountETHMin,
         address _to,
-        uint256 _deadline
-    )
-        external
-        payable
-        returns (
-            uint256 _amountToken,
-            uint256 _amountETH,
-            uint256 _liquidity
-        );
+        uint _deadline
+    ) external payable returns (uint _amountToken, uint _amountETH, uint _liquidity);
 
     function removeLiquidity(
         address _tokenA,
         address _tokenB,
-        uint256 _liquidity,
-        uint256 _amountAMin,
-        uint256 _amountBMin,
+        uint _liquidity,
+        uint _amountAMin,
+        uint _amountBMin,
         address _to,
-        uint256 _deadline
-    ) external returns (uint256 _amountA, uint256 _amountB);
+        uint _deadline
+    ) external returns (uint _amountA, uint _amountB);
 
     function removeLiquidityETH(
         address _token,
-        uint256 _liquidity,
-        uint256 _amountTokenMin,
-        uint256 _amountETHMin,
+        uint _liquidity,
+        uint _amountTokenMin,
+        uint _amountETHMin,
         address _to,
-        uint256 _deadline
-    ) external returns (uint256 _amountToken, uint256 _amountETH);
+        uint _deadline
+    ) external returns (uint _amountToken, uint _amountETH);
 
     function swapExactTokensForTokens(
-        uint256 _amountIn,
-        uint256 _amountOutMin,
+        uint _amountIn,
+        uint _amountOutMin,
         address[] calldata _path,
         address _to,
-        uint256 _deadline
-    ) external returns (uint256[] memory _amounts);
+        uint _deadline
+    ) external returns (uint[] memory _amounts);
 
     function swapTokensForExactTokens(
-        uint256 _amountOut,
-        uint256 _amountInMax,
+        uint _amountOut,
+        uint _amountInMax,
         address[] calldata _path,
         address _to,
-        uint256 _deadline
-    ) external returns (uint256[] memory _amounts);
+        uint _deadline
+    ) external returns (uint[] memory _amounts);
 
     function swapExactETHForTokens(
-        uint256 _amountOutMin,
+        uint _amountOutMin,
         address[] calldata _path,
         address _to,
-        uint256 _deadline
-    ) external payable returns (uint256[] memory _amounts);
+        uint _deadline
+    ) external payable returns (uint[] memory _amounts);
 
     function swapTokensForExactETH(
-        uint256 _amountOut,
-        uint256 _amountInMax,
+        uint _amountOut,
+        uint _amountInMax,
         address[] calldata _path,
         address _to,
-        uint256 _deadline
-    ) external returns (uint256[] memory _amounts);
+        uint _deadline
+    ) external returns (uint[] memory _amounts);
 
     function swapExactTokensForETH(
-        uint256 _amountIn,
-        uint256 _amountOutMin,
+        uint _amountIn,
+        uint _amountOutMin,
         address[] calldata _path,
         address _to,
-        uint256 _deadline
-    ) external returns (uint256[] memory amounts);
+        uint _deadline
+    ) external returns (uint[] memory amounts);
 
     function swapETHForExactTokens(
-        uint256 _amountOut,
+        uint _amountOut,
         address[] calldata _path,
         address _to,
-        uint256 _deadline
-    ) external payable returns (uint256[] memory amounts);
+        uint _deadline
+    ) external payable returns (uint[] memory amounts);
 }
