+++ contracts/tokens/interfaces/IWETH.sol
@@ -4,8 +4,6 @@ pragma solidity ^0.8.10;
 
 interface IWETH {
     function deposit() external payable;
-
-    function transfer(address to, uint256 value) external returns (bool);
-
-    function withdraw(uint256) external;
+    function transfer(address to, uint value) external returns (bool);
+    function withdraw(uint) external;
 }
